:sparkles: State: Extracting Form Logic
