+++ src/components/BlogPostForm.js
@@ -1,0 +1,37 @@
+import React, { useState } from 'react';
+import { View, Text, StyleSheet, TextInput, Button } from 'react-native';
+
+const BlogPostForm = (props) => {
+  const { onSubmit, initialValues, labels } = props;
+
+  const [title, setTitle] = useState('');
+  const [content, setContent] = useState('');
+
+  return (
+    <View>
+      <Text style={styles.label}>Enter Title:</Text>
+      <TextInput style={styles.input} value={title} onChangeText={setTitle}/>
+      <Text style={styles.label}>Enter Content:</Text>
+      <TextInput style={styles.input} value={content} onChangeText={setContent} />
+      <Button title="Save Blog Post" onPress={() => {}} />
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  input: {
+    fontSize: 18,
+    borderWidth: 1,
+    borderColor: 'black',
+    marginBottom: 15,
+    padding: 5,
+    margin: 5
+  },
+  label: {
+    fontSize: 20,
+    marginBottom: 5,
+    marginLeft: 5
+  }
+});
+
+export default BlogPostForm;
