+++ src/screens/EditScreen.js
@@ -1,7 +1,8 @@
-import React, { useState, useContext } from 'react';
-import { View, Text, StyleSheet, TextInput } from 'react-native';
+import React, { useContext } from 'react';
+import { StyleSheet } from 'react-native';
 
 import { Context as BlogContext } from "../context/BlogContext";
+import BlogPostForm from "../components/BlogPostForm";
 
 const EditScreen = (props) => {
   const { navigation } = props;
@@ -9,17 +10,7 @@ const EditScreen = (props) => {
   const { state } = useContext(BlogContext);
   const blogPost = state.find(x => x.id === id);
 
-  const [title, setTitle] = useState(blogPost.title);
-  const [content, setContent] = useState(blogPost.content);
-
-  return (
-    <View>
-      <Text>Edit Title:</Text>
-      <TextInput value={title} onChangeText={setTitle} />
-      <Text>Edit Content:</Text>
-      <TextInput value={content} onChangeText={setContent} />
-    </View>
-  );
+  return <BlogPostForm />;
 };
 
 const styles = StyleSheet.create({});
